refactor(vc): extract vc_issuance.go from vc_service.go
Pure move of execution/workflow VC issuance + signing logic into its
own file within the same services package. No behavior change.

Moves: GenerateExecutionVC, CreateWorkflowVC, createVCDocument, signVC,
SignAgentTagVC, generateWorkflowVCDocument, createWorkflowVCDocument,
signWorkflowVC, determineWorkflowStatus, countCompletedSteps.

Refs #415
